backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
chess>=1.10.0
anthropic>=0.28.0
httpx>=0.27.0
python-multipart>=0.0.9
redis>=5.0
psycopg2-binary>=2.9
websockets>=11
# stockfish binary must be installed at the OS level (e.g. apt install stockfish)
